Address Philip's comments
