+++ src/metatrain/utils/data/dataset.py
@@ -14,10 +14,17 @@ from metatomic.torch import load_system
 from omegaconf import DictConfig
 from torch.utils.data import Subset
 
-from ..external_naming import to_external_name
-from ..units import get_gradient_units
-from .readers.metatensor import _check_tensor_map_metadata, _empty_tensor_map_like
-from .target_info import TargetInfo, get_energy_target_info, get_generic_target_info
+from metatrain.utils.data.readers.metatensor import (
+    _check_tensor_map_metadata,
+    _empty_tensor_map_like,
+)
+from metatrain.utils.data.target_info import (
+    TargetInfo,
+    get_energy_target_info,
+    get_generic_target_info,
+)
+from metatrain.utils.external_naming import to_external_name
+from metatrain.utils.units import get_gradient_units
 
 
 class DatasetInfo:
@@ -452,42 +459,6 @@ def _is_disk_dataset(dataset: Any) -> bool:
     if isinstance(dataset, torch.utils.data.Subset):
         return _is_disk_dataset(dataset.dataset)
     return False
-
-
-# class DiskDatasetWriter:
-#     """
-#     A class for writing a dataset to disk, to be read by the :py:class:`DiskDataset`
-#     class.
-
-#     The class is initialized with a path to a zip file, and samples can be written to
-#     the zip file using the :py:meth:`write_sample` method.
-
-#     :param path: Path to the zip file to write the dataset to.
-#     """
-
-#     def __init__(self, path: Union[str, Path], mode: Literal["w", "a"] = "w"):
-#         self.zip_file = zipfile.ZipFile(path, mode)
-#         self.index = 0
-
-#     def write_sample(self, system: System, targets: Dict[str, TensorMap]):
-#         """
-#         Write a sample to the zip file.
-
-#         :param system: The system to write.
-#         :param targets: A dictionary of targets to write, where each value is
-#             a :py:class:`TensorMap`.
-#         """
-#         with self.zip_file.open(f"{self.index}/system.mta", "w") as file:
-#             mta.save(file, system)
-#         for target_name, target in targets.items():
-#             with self.zip_file.open(f"{self.index}/{target_name}.mts", "w") as file:
-#                 tensor_buffer = mts.save_buffer(target)
-#                 numpy_buffer = tensor_buffer.numpy()
-#                 np.save(file, numpy_buffer)
-#         self.index += 1
-
-#     def close(self):
-#         self.zip_file.close()
 
 
 def _save_indices(
